Split Operations.js into src/operations/{factories.ts, sequences.js, aggregations.js, helpers.js}
Operations.js is monolithic and creates an init-time circular dependency
that prevents Collection.ts from importing it: `Operations.js → Map.js →
OrderedMap.js → List.js → class ListImpl extends IndexedCollectionImpl`
crashes when Collection.ts is mid-load. Even sorting out Map/OrderedMap
doesn't fix it because Operations.js itself has `class ToKeyedSequence
extends KeyedSeqImpl` at the top level, which fails when Seq.js is
mid-load.

Split into 4 files in src/operations/:
- helpers.js (JS): reify, makeSequence, collectionClass,
  defaultComparator, cacheResultThrough. No class extends, no Map import,
  safe to load from anywhere.
- factories.ts (TS): the 14 pure factories (flipFactory, mapFactory,
  reverseFactory, filterFactory, sliceFactory, takeWhileFactory,
  skipWhileFactory, flattenFactory, flatMapFactory, interposeFactory,
  sortFactory, maxFactory, zipWithFactory, partitionFactory) + local
  maxCompare. Pragmatic typing (`any` for polymorphic params) since the
  strict public API types live in type-definitions/immutable.d.ts.
- sequences.js (JS): the 5 Sequence classes (ToKeyedSequence,
  ToIndexedSequence, ToSetSequence, FromEntriesSequence, ConcatSeq) +
  concatFactory + validateEntry + prototype assignments.
- aggregations.js (JS): countByFactory, groupByFactory (the only two
  consumers of Map/OrderedMap).

Update direct imports in Map.js, Set.js (sortFactory) and
CollectionImpl.js to point at the precise sub-files instead of the old
single Operations.js. No index.ts: an index that re-exports everything
would re-create the same load-time cascade.

Adds `dpdm` and the `check-circular-deps` npm script for verification.

Result: 14 → 12 circular dependencies. The two `Map.js → Operations.js`
cycles are gone. The remaining cycles are unrelated (Map ↔ methods/*
↔ functional/*, OrderedSet → CollectionImpl, Seq.js → Collection.ts).
All 746 tests still pass.
